fix: disable save button when communication from is open
